Program #1(a)
Description: Given a height (in feet) entered by the user, this program
determine how long (in seconds) it will take an object dropped from that
height to hit the ground, using the kinematic equation t = sqrt(height /
(0.5 * 32.174)), where 32.174 is the average acceleration due to gravity
on Earth. In addition, it will report the object's speed (s= 32.174 * t)
at the time of impact.